refactor: optimize blockchain event indexing
- Extract common indexer logic into runIndexers function
- Implement chunked processing for EVM networks only
- Keep single-range processing for Tron networks
- Remove duplicate code and improve maintainability

This change improves the code by:
- Separating network-specific block processing logic
- Reducing code duplication
- Making the indexing process more efficient
- Maintaining different processing strategies for EVM and Tron
