+++ services/common/order.go
@@ -21,10 +21,8 @@ import (
 	"github.com/paycrest/aggregator/ent/providerprofile"
 	"github.com/paycrest/aggregator/ent/provisionbucket"
 	"github.com/paycrest/aggregator/ent/receiveaddress"
-	"github.com/paycrest/aggregator/ent/senderprofile"
 	tokenent "github.com/paycrest/aggregator/ent/token"
 	"github.com/paycrest/aggregator/ent/transactionlog"
-	"github.com/paycrest/aggregator/ent/user"
 	db "github.com/paycrest/aggregator/storage"
 	"github.com/paycrest/aggregator/types"
 	"github.com/paycrest/aggregator/utils"
@@ -818,152 +816,6 @@ func CheckAMLCompliance(rpcUrl string, txHash string) (bool, error) {
 
 	if backfillTransaction.Decision == "Allow" {
 		return true, nil
-	}
-
-	return false, nil
-}
-
-// UpdateReceiveAddressStatus updates the status of a receive address based on a transfer event.
-func UpdateReceiveAddressStatus(
-	ctx context.Context,
-	receiveAddress *ent.ReceiveAddress,
-	paymentOrder *ent.PaymentOrder,
-	event *types.TokenTransferEvent,
-	createOrder func(ctx context.Context, orderID uuid.UUID) error,
-	getProviderRate func(ctx context.Context, providerProfile *ent.ProviderProfile, tokenSymbol string, currency string) (decimal.Decimal, error),
-) (done bool, err error) {
-	if event.To == receiveAddress.Address {
-		// Check for existing address with txHash
-		count, err := db.Client.ReceiveAddress.
-			Query().
-			Where(receiveaddress.TxHashEQ(event.TxHash)).
-			Count(ctx)
-		if err != nil {
-			return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-		}
-
-		if count > 0 && receiveAddress.Status != receiveaddress.StatusUnused {
-			// This transfer has already been indexed
-			return false, nil
-		}
-
-		// Check for existing payment order with txHash
-		if paymentOrder.TxHash == event.TxHash {
-			// This transfer has already been indexed
-			return false, nil
-		}
-
-		// This is a transfer to the receive address to create an order on-chain
-		// Compare the transferred value with the expected order amount + fees
-		fees := paymentOrder.NetworkFee.Add(paymentOrder.SenderFee).Add(paymentOrder.ProtocolFee)
-		orderAmountWithFees := paymentOrder.Amount.Add(fees).Round(int32(paymentOrder.Edges.Token.Decimals))
-		transferMatchesOrderAmount := event.Value.Equal(orderAmountWithFees)
-
-		tx, err := db.Client.Tx(ctx)
-		if err != nil {
-			return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-		}
-
-		paymentOrderUpdate := tx.PaymentOrder.Update().Where(paymentorder.IDEQ(paymentOrder.ID))
-		if paymentOrder.ReturnAddress == "" {
-			paymentOrderUpdate = paymentOrderUpdate.SetReturnAddress(event.From)
-		}
-
-		orderRecipient := paymentOrder.Edges.Recipient
-		if !transferMatchesOrderAmount {
-			// Update the order amount will be updated to whatever amount was sent to the receive address
-			newOrderAmount := event.Value.Sub(fees.Round(int32(paymentOrder.Edges.Token.Decimals)))
-			paymentOrderUpdate = paymentOrderUpdate.SetAmount(newOrderAmount.Round(int32(paymentOrder.Edges.Token.Decimals)))
-			// Update the rate with the current rate if order is older than 30 mins for a P2P order from the sender dashboard
-			if strings.HasPrefix(orderRecipient.Memo, "P#P") && orderRecipient.ProviderID != "" && paymentOrder.CreatedAt.Before(time.Now().Add(-30*time.Minute)) {
-				providerProfile, err := db.Client.ProviderProfile.
-					Query().
-					Where(
-						providerprofile.HasUserWith(
-							user.HasSenderProfileWith(
-								senderprofile.HasPaymentOrdersWith(
-									paymentorder.IDEQ(paymentOrder.ID),
-								),
-							),
-						),
-					).
-					Only(ctx)
-				if err != nil {
-					return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-				}
-
-				institution, err := utils.GetInstitutionByCode(ctx, orderRecipient.Institution, true)
-				if err != nil {
-					return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-				}
-
-				rate, err := getProviderRate(ctx, providerProfile, paymentOrder.Edges.Token.Symbol, institution.Edges.FiatCurrency.Code)
-				if err != nil {
-					return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-				}
-				paymentOrderUpdate = paymentOrderUpdate.SetRate(rate)
-			}
-			transferMatchesOrderAmount = true
-		}
-
-		if paymentOrder.AmountPaid.GreaterThanOrEqual(decimal.Zero) && paymentOrder.AmountPaid.LessThan(orderAmountWithFees) {
-			transactionLog, err := tx.TransactionLog.
-				Create().
-				SetStatus(transactionlog.StatusCryptoDeposited).
-				SetGatewayID(paymentOrder.GatewayID).
-				SetTxHash(event.TxHash).
-				SetNetwork(paymentOrder.Edges.Token.Edges.Network.Identifier).
-				SetMetadata(map[string]interface{}{
-					"GatewayID":       paymentOrder.GatewayID,
-					"transactionData": event,
-				}).
-				Save(ctx)
-			if err != nil {
-				return true, fmt.Errorf("UpdateReceiveAddressStatus.transactionlog: %v", err)
-			}
-
-			_, err = paymentOrderUpdate.
-				SetFromAddress(event.From).
-				SetTxHash(event.TxHash).
-				SetBlockNumber(int64(event.BlockNumber)).
-				AddAmountPaid(event.Value).
-				AddTransactions(transactionLog).
-				Save(ctx)
-			if err != nil {
-				return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-			}
-
-			// Commit the transaction
-			if err := tx.Commit(); err != nil {
-				return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-			}
-		}
-
-		if transferMatchesOrderAmount {
-			// Transfer value equals order amount with fees
-			_, err = receiveAddress.
-				Update().
-				SetStatus(receiveaddress.StatusUsed).
-				SetLastUsed(time.Now()).
-				SetTxHash(event.TxHash).
-				SetLastIndexedBlock(int64(event.BlockNumber)).
-				Save(ctx)
-			if err != nil {
-				return true, fmt.Errorf("UpdateReceiveAddressStatus.db: %v", err)
-			}
-
-			err = createOrder(ctx, paymentOrder.ID)
-			if err != nil {
-				return true, fmt.Errorf("UpdateReceiveAddressStatus.CreateOrder: %v", err)
-			}
-
-			return true, nil
-		}
-
-		err = HandleReceiveAddressValidity(ctx, receiveAddress, paymentOrder)
-		if err != nil {
-			return true, fmt.Errorf("UpdateReceiveAddressStatus.HandleReceiveAddressValidity: %v", err)
-		}
 	}
 
 	return false, nil
